requirements.txt
dnspython==1.14.0
requests==2.10.0
